Update to 4.3

+++ src/editor/import_gif_to_animated_texture.cpp
@@ -160,7 +160,7 @@ String GifToAnimatedTextureImportPlugin::_get_resource_type() const
     return "AnimatedTexture";
 }
 
-double GifToAnimatedTextureImportPlugin::_get_priority() const
+float GifToAnimatedTextureImportPlugin::_get_priority() const
 {
     return 1.0;
 }
